+++ src/utils.jl
@@ -1,99 +1,71 @@
 """
     get_var_capex(m, prefix::Symbol)
-
-Extracts the CAPEX variable with a given `prefix` from the model.
-"""
-get_var_capex(m, prefix::Symbol) = m[Symbol(prefix, :_capex)]
-"""
     get_var_capex(m, prefix::Symbol, element)
 
-When the type `element` is used as conditional input, it extracts only the variable for
-the specified element.
+Extracts the CAPEX variable with a given `prefix` from the model or only the variable for
+    the specified `element`.
 """
+get_var_capex(m, prefix::Symbol) = m[Symbol(prefix, :_capex)]
 get_var_capex(m, prefix::Symbol, element) = m[Symbol(prefix, :_capex)][element, :]
 
 """
     get_var_inst(m, prefix::Symbol)
-
-Extracts the installed capacity variable with a given `prefix` from the model.
-"""
-get_var_inst(m, prefix::Symbol) = m[Symbol(prefix, :_inst)]
-"""
     get_var_inst(m, prefix::Symbol, element)
 
-When the type `element` is used as conditional input, it extracts only the variable for
-the specified element.
+Extracts the installed capacity variable with a given `prefix` from the model or only the
+variable for the specified `element`.
 """
+get_var_inst(m, prefix::Symbol) = m[Symbol(prefix, :_inst)]
 get_var_inst(m, prefix::Symbol, element) = m[Symbol(prefix, :_inst)][element, :]
 
 """
     get_var_current(m, prefix::Symbol)
-
-Extracts the current capacity variable with a given `prefix` from the model.
-"""
-get_var_current(m, prefix::Symbol) = m[Symbol(prefix, :_current)]
-"""
     get_var_current(m, prefix::Symbol, element)
 
-When the type `element` is used as conditional input, it extracts only the variable for
-the specified element.
+Extracts the current capacity variable with a given `prefix` from the model or only the
+variable for the specified `element`.
 """
+get_var_current(m, prefix::Symbol) = m[Symbol(prefix, :_current)]
 get_var_current(m, prefix::Symbol, element) = m[Symbol(prefix, :_current)][element, :]
 
 """
     get_var_add(m, prefix::Symbol)
-
-Extracts the investment capacity variable with a given `prefix` from the model.
-"""
-get_var_add(m, prefix::Symbol) = m[Symbol(prefix, :_add)]
-"""
     get_var_add(m, prefix::Symbol, element)
 
-When the type `element` is used as conditional input, it extracts only the variable for
-the specified element.
+Extracts the investment capacity variable with a given `prefix` from the model or only the
+variable for the specified `element`.
 """
+get_var_add(m, prefix::Symbol) = m[Symbol(prefix, :_add)]
 get_var_add(m, prefix::Symbol, element) = m[Symbol(prefix, :_add)][element, :]
 
 """
     get_var_rem(m, prefix::Symbol)
-
-Extracts the retired capacity variable with a given `prefix` from the model.
-"""
-get_var_rem(m, prefix::Symbol) = m[Symbol(prefix, :_rem)]
-"""
     get_var_rem(m, prefix::Symbol, element)
 
-When the type `element` is used as conditional input, it extracts only the variable for
-the specified element.
+Extracts the retired capacity variable with a given `prefix` from the model or only the
+variable for the specified `element`.
 """
+get_var_rem(m, prefix::Symbol) = m[Symbol(prefix, :_rem)]
 get_var_rem(m, prefix::Symbol, element) = m[Symbol(prefix, :_rem)][element, :]
 
 """
     get_var_invest_b(m, prefix::Symbol)
-
-Extracts the binary investment variable with a given `prefix` from the model.
-"""
-get_var_invest_b(m, prefix::Symbol) = m[Symbol(prefix, :_invest_b)]
-"""
     get_var_invest_b(m, prefix::Symbol, element)
 
-When the type `element` is used as conditional input, it extracts only the variable for
-the specified element.
+Extracts the binary investment variable with a given `prefix` from the model or only the
+variable for the specified `element`.
 """
+get_var_invest_b(m, prefix::Symbol) = m[Symbol(prefix, :_invest_b)]
 get_var_invest_b(m, prefix::Symbol, element) = m[Symbol(prefix, :_invest_b)][element, :]
 
 """
     get_var_remove_b(m, prefix::Symbol)
-
-Extracts the binary retirement variable with a given `prefix` from the model.
-"""
-get_var_remove_b(m, prefix::Symbol) = m[Symbol(prefix, :_remove_b)]
-"""
     get_var_remove_b(m, prefix::Symbol, element)
 
-When the type `element` is used as conditional input, it extracts only the variable for
-the specified element.
+Extracts the binary retirement variable with a given `prefix` from the model or only the
+variable for the specified `element`.
 """
+get_var_remove_b(m, prefix::Symbol) = m[Symbol(prefix, :_remove_b)]
 get_var_remove_b(m, prefix::Symbol, element) = m[Symbol(prefix, :_remove_b)][element, :]
 
 """
@@ -117,7 +89,8 @@ set_capex_value(m, element, inv_data, prefix, 𝒯ᴵⁿᵛ) =
     set_capex_value(m, element, inv_data, prefix, 𝒯ᴵⁿᵛ, ::Investment)
 
 When no specialized method is defined for the investment mode, it calculates the capital
-cost based on the multiplication of the field `capex` in `inv_data` with the added capacity.
+cost based on the multiplication of the field `capex` in `inv_data` with the added capacity
+extracted from the model through the function [`get_var_add`](@ref).
 """
 function set_capex_value(m, element, inv_data, prefix, 𝒯ᴵⁿᵛ, ::Investment)
     # Deduce the required variable